pyproject.toml
[project]
name = "localization"
version = "0.0.0"
description = "Localization and addressing service for the MGO project"
authors = [{ name = "Ministerie van Volksgezondheid, Welzijn en Sport" }]
requires-python = ">=3.11,<4"
readme = "README.md"
license = "EUPL-1.2"
dependencies = [
    "uvicorn>=0.41.0,<0.42",
    "fastapi>=0.131.0,<0.136",
    "pydantic-settings>=2.12.0,<3",
    "sqlalchemy>=2.0.45,<3",
    "psycopg[binary, pool]>=3.3.2,<4",
    "inject>=5.3.0,<6",
    "xmltodict>=1.0.4,<2",
    "requests>=2.32.5,<3",
    "fhir-resources>=8.1.0,<9",
    "cryptography>=46.0.5,<47",
    "orjson>=3.11.5,<4",
    "defusedxml>=0.7.1,<0.8",
    "lxml>=6.0.2,<7",
    "pyproj>=3.7.2,<4",
    "brotli>=1.1.0,<2",
    "jwcrypto>=1.5.6,<2",
    "types-jwcrypto>=1.5.0.20251102,<2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2,<10",
    "pytest-cov>=7.0.0,<8",
    "httpx>=0.28.1,<0.29",
    "ruff>=0.15.2,<0.16",
    "codespell>=2.4.1,<3",
    "faker>=40.1.2,<41",
    "pytest-mock>=3.15.1,<4",
    "types-requests>=2.32.4,<3",
    "mypy>=1.19.1,<2",
    "types-defusedxml>=0.7.0.20250822,<0.8",
    "freezegun>=1.5.5,<2",
    "pytest-watcher>=0.6.3,<0.7",
    "debugpy>=1.8.19,<2",
    "pytest-asyncio>=1.3.0,<2",
    "black>=26.1.0,<27",
    "pre-commit>=4.5.1",
    "pre-commit-uv>=4.2.1",
]

[tool.uv]
default-groups = "all"
package = false

[build-system]
requires = ["uv_build>=0.9.30,<0.10.0"]
build-backend = "uv_build"

[tool.ruff]
cache-dir = "~/.cache/ruff"
include = [
    "pyproject.toml",
    "app/*.py",
    "tests/*.py",
]
line-length = 120

[tool.black]
line-length = 120

[tool.ruff.lint]
select = ["E", "ERA", "F", "B", "SIM", "I", "TID251", "TID252"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"unittest".msg = "use pytest-mock instead"

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.params.Depends", "fastapi.Query", "fastapi.params.Query", "app.utils.resolve_instance"]

[tool.mypy]
files = "app,tests"
python_version = "3.12"
cache_dir = "~/.cache/mypy"
plugins = ['pydantic.mypy']
strict = true
disallow_any_unimported = true
disallow_any_explicit = true

[tool.pydantic-mypy]
init_forbid_extra = true
init_typed = true

[tool.pytest.ini_options]
cache_dir = "~/.cache/pytest"
testpaths = [
    "tests",
]

[tool.codespell]
skip = [
    "./static/**",
    "./secrets/*",
    "./data/*",
    "./storage/**",
]
